extension AccessPhoto: UIImagePickerControllerDelegate 成功運作“點擊圖片＞上傳至資料庫＞取得圖片網址
