+++ mapUFirestore/AccessPhoto.swift
@@ -11,7 +11,7 @@ import FirebaseFirestoreSwift
 import FirebaseStorage
 import FirebaseAuth
 
-class AccessPhoto: UIViewController {
+class AccessPhoto: UIViewController, UINavigationControllerDelegate {
 
     let imagePickerController = UIImagePickerController()
     let storageRef = Storage.storage().reference()
@@ -54,58 +54,21 @@ class AccessPhoto: UIViewController {
     
         }
     
+
     
-//    彼得潘的上傳檔案範例，但會報錯閃退
-//    func uploadSampleFromPeterPan() {
-//        let fileReference = self.storageRef.child("001.jpeg")
-//
-//        let image = UIImage(named: "001")
-//        let imageData = image?.jpegData(compressionQuality: 0.9)
-//
-//        if let data = try? Data(contentsOf: Bundle.main.url(forResource: "001", withExtension: "jpeg")!) {
-//           fileReference.putData(data, metadata: nil) { (metadata, error) in
-//              guard let _ = metadata, error == nil else {
-//                 print("no metadata")
-//                 return
-//              }
-//              fileReference.downloadURL(completion: { (url, error) in
-//                 guard let downloadURL = url else {
-//                    print("no URL")
-//                    return
-//                 }
-//                 print(downloadURL)
-//              })
-//           }
-//        }
-//    }
-    
-    
-//    彼得潘的sample code:上傳檔案 & 將檔案路徑存在 database
-//    func uploadPhoto(completion: @escaping (URL?) -> () ) {
-//
-//            let fileReference = Storage.storage().reference().child(UUID().uuidString + ".jpg")
-//
-//
-//            if let data = photoButton.image(for: .normal)?.jpegData(compressionQuality: 0.9) {
-//                fileReference.putData(data, metadata: nil) { (_, error) in
-//                    guard error == nil else {
-//                        print("upload error")
-//                        return
-//                    }
-//                    fileReference.downloadURL { (url, error) in
-//                        completion(url)
-//                    }
-//                }
-//            }
-//    }
+
     
     
     @IBAction func DennyTeacherUploadPic(_ sender: UIButton) {
-        
+//        實例化一個選擇asset的class
         let imagePicker = UIImagePickerController()
+//        asset來源設定為裝置內所貯存的asset
         imagePicker.sourceType = .photoLibrary
+//        設定為不可編輯此asset
         imagePicker.allowsEditing = false
+//        顯示圖素之處設定為self，這個寫法比還待確認其意義
         imagePicker.delegate = self
+//        跳出選擇asset的彈窗
         present(imagePicker, animated: true, completion: nil)
     }
     
@@ -128,7 +91,7 @@ class AccessPhoto: UIViewController {
     
     func theSrouceToAccessPhoto() {
         
-        let controller = UIAlertController(title: "拍照?從照片選取?從相簿選取?", message: "", preferredStyle: .alert)
+        let controller = UIAlertController(title: "修改頭圖", message: "", preferredStyle: .alert)
             controller.view.tintColor = UIColor.gray
 
             // 相機
@@ -267,68 +230,91 @@ class AccessPhoto: UIViewController {
 extension AccessPhoto: UIImagePickerControllerDelegate {
     public func imagePickerController(_ picker: UIImagePickerController, didFinishPickingMediaWithInfo info: [UIImagePickerController.InfoKey : Any]) {
         
-        // info 用來取得不同類型的圖片，此 Demo 的型態為 originaImage，其它型態有影片、修改過的圖片等等
+        // info 用來取得不同類型的圖片，此 Demo 的型態為 originaImage，其它型態有影片、修改過的圖片等等，這段就是執行“點擊圖片之後，顯示在頁面上”的 method
         if let image = info[.originalImage] as? UIImage {
-            self.photoImageView.image = image
-        }
-        
-        picker.dismiss(animated: true)
-    }
-}
-
-extension AccessPhoto: UINavigationControllerDelegate { }
-
-extension PhotoCollectionView: UIImagePickerControllerDelegate, UINavigationControllerDelegate {
-        
-        func imagePickerController(_ picker: UIImagePickerController, didFinishPickingMediaWithInfo info: [String : Any]) {
             
-            var selectedImageFromPicker: UIImage?
-            
-            // 取得從 UIImagePickerController 選擇的檔案
-            if let pickedImage = info[UIImagePickerController.InfoKey.originalImage.rawValue] as? UIImage {
+            let fileReference = Storage.storage().reference().child(UUID().uuidString + ".jpg")
                 
-                selectedImageFromPicker = pickedImage
+            if let imageData = image.jpegData(compressionQuality: 0.9) {
+                     fileReference.putData(imageData, metadata: nil) { (_, error) in
+                         guard error == nil else {
+                             print("upload error")
+                             return
+                         }
+                                               
+                        fileReference.downloadURL (completion: { (url, error) in
+                            guard let downloadURL = url else {
+                               print("no URL")
+                               return
+                            }
+                            print("this is \(downloadURL)")
+                             
+                        })
+                     }
+                 }
             }
             
-            // 可以自動產生一組獨一無二的 ID 號碼，方便等一下上傳圖片的命名
-            let uniqueString = NSUUID().uuidString
-            
-            // 當判斷有 selectedImage 時，我們會在 if 判斷式裡將圖片上傳
-            if let selectedImage = selectedImageFromPicker {
-                
-                let storageRef = Storage.storage().reference().child("AppCodaFireUpload").child("\(uniqueString).png")
-                    
-                if let uploadData = selectedImage.pngData() {
-                        // 這行就是 FirebaseStorage 關鍵的存取方法。
-                    
-                    storageRef.putData(uploadData, metadata: nil, completion: { (data, error) in
-                            
-                            if error != nil {
-                                
-                                // 若有接收到錯誤，我們就直接印在 Console 就好，在這邊就不另外做處理。
-                                print("Error: \(error!.localizedDescription)")
-                                return
-                            }
-                            
-                            // 連結取得方式
-                         storageRef.downloadURL{ (url, error) in
-                            
-                            guard let downloadURL = url else {
-                                print("Failed to get URL!")
-                                return
-                            }
-                   // 我們可以 print 出來看看這個連結事不是我們剛剛所上傳的照片。
-                            print("Photo Url: \(downloadURL)")
-                            
-                            
-                        }
-                        
-                        })
-                    }            }
-            
-            dismiss(animated: true, completion: nil)
+//            self.photoImageView.image = image
+        picker.dismiss(animated: true)
         }
-    
-    
-    
+        
+     
     }
+
+
+//extension AccessPhoto: UINavigationControllerDelegate { }
+
+//extension PhotoCollectionView: UIImagePickerControllerDelegate, UINavigationControllerDelegate {
+//        
+//        func imagePickerController(_ picker: UIImagePickerController, didFinishPickingMediaWithInfo info: [String : Any]) {
+//            
+//            var selectedImageFromPicker: UIImage?
+//            
+//            // 取得從 UIImagePickerController 選擇的檔案
+//            if let pickedImage = info[UIImagePickerController.InfoKey.originalImage.rawValue] as? UIImage {
+//                
+//                selectedImageFromPicker = pickedImage
+//            }
+//            
+//            // 可以自動產生一組獨一無二的 ID 號碼，方便等一下上傳圖片的命名
+//            let uniqueString = NSUUID().uuidString
+//            
+//            // 當判斷有 selectedImage 時，我們會在 if 判斷式裡將圖片上傳
+//            if let selectedImage = selectedImageFromPicker {
+//                
+//                let storageRef = Storage.storage().reference().child("AppCodaFireUpload").child("\(uniqueString).png")
+//                    
+//                if let uploadData = selectedImage.pngData() {
+//                        // 這行就是 FirebaseStorage 關鍵的存取方法。
+//                    
+//                    storageRef.putData(uploadData, metadata: nil, completion: { (data, error) in
+//                            
+//                            if error != nil {
+//                                
+//                                // 若有接收到錯誤，我們就直接印在 Console 就好，在這邊就不另外做處理。
+//                                print("Error: \(error!.localizedDescription)")
+//                                return
+//                            }
+//                            
+//                            // 連結取得方式
+//                         storageRef.downloadURL{ (url, error) in
+//                            
+//                            guard let downloadURL = url else {
+//                                print("Failed to get URL!")
+//                                return
+//                            }
+//                   // 我們可以 print 出來看看這個連結事不是我們剛剛所上傳的照片。
+//                            print("Photo Url: \(downloadURL)")
+//                            
+//                            
+//                        }
+//                        
+//                        })
+//                    }            }
+//            
+//            dismiss(animated: true, completion: nil)
+//        }
+//    
+//    
+//    
+//    }
